+++ opencog/atoms/reduct/PutLink.cc
@@ -22,6 +22,7 @@
 
 #include <opencog/atomspace/atom_types.h>
 #include <opencog/atomspace/ClassServer.h>
+#include <opencog/atoms/execution/Instantiator.h>
 #include "PutLink.h"
 
 using namespace opencog;
