+++ opencog/query/DefaultPatternMatchCB.cc
@@ -22,11 +22,11 @@
  */
 
 #include <opencog/atoms/execution/EvaluationLink.h>
+#include <opencog/atoms/execution/Instantiator.h>
 #include <opencog/atomutils/FindUtils.h>
 #include <opencog/atoms/bind/BetaRedex.h>
 
 #include "DefaultPatternMatchCB.h"
-#include "Instantiator.h"
 
 using namespace opencog;
